fix(swiper): fix bug in quick reverse direction in first page
BUG: In first page, first up, then quick reverse to down, the page may stuck in moving phase.

REASON: when quick move, when active page is not EMPTY_PAGE, the last frame may rendered correctly, but in next trigger of moving event, the activePage may be EMPTY_PAGE, and offset may be a negative value. In the original logic, when activePage is EMTPY_PAGE, it will not render. so the page was stuck.

FIX:
1. Let page rendered when activePage is EMTPY
2. Reset offset to 0 when activePage is EMPTY
3. To avoid misjudgment of ending status, we set pageChange value more accurate when activePage is EMPTY